TransformationChain: Forbid copy constructor
Copying the TransformationChain is not possible currently. It stores
references to classes of a base type (wrapped into a smart pointer), so
the runtime can't determine the size of each object stored. Cloning is
no standard functionality in C++.
Copying a TransformationChain even fails because unique pointers do not
allow a copy, only move.